# Add AVX2 `polyvec_{de,}compress`