@mkannwischer 2024-11-19
After the recent merges (adding PMU support, adding LTO support), let's re-do the benchmarks with this PR rebased on top of 75f52dc1c259d69bb4e30ef86886a9c0e01c5935. 

**TL;LR: We do see performance gains for encaps and decaps of up to 8% with this PR. This is consistent accross platforms (Raptor Lake, c7i) and compiler versions (gcc 13.2.0, gcc 14.2.1). `polyvec_compress` is 4-7x faster.  `polyvec_decompress` is 5-11x faster.** 

Here are the results on my Raptor Lake (gcc 14.2.1)
main (75f52dc1c259d69bb4e30ef86886a9c0e01c5935)
```
./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 13101
    encaps cycles = 15894
    decaps cycles = 21068

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  13021  13054  13068  13079  13090  13101  13114  13130  13154  13211  28166
    encaps percentiles:  15820  15851  15864  15873  15885  15894  15910  15934  15967  16039  35007
    decaps percentiles:  20974  20999  21019  21036  21052  21068  21103  21157  21221  21320  44787

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 22231
    encaps cycles = 24024
    decaps cycles = 31583

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  22143  22182  22197  22208  22220  22231  22246  22265  22308  22830  23650
    encaps percentiles:  23932  23975  23990  24001  24011  24024  24041  24055  24096  24632  25470
    decaps percentiles:  31489  31528  31543  31557  31571  31583  31598  31619  31662  32220  33021

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 31818
    encaps cycles = 35352
    decaps cycles = 45531

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  31671  31725  31753  31773  31796  31818  31843  31882  31990  33067  34430
    encaps percentiles:  35195  35257  35290  35310  35336  35352  35380  35421  35523  36596  37939
    decaps percentiles:  45350  45407  45444  45467  45494  45531  45595  45689  45915  46762  49307
```

this PR (056112032b689bfa59322fdff8209390fa48736a):
```
./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 13105
    encaps cycles = 15289
    decaps cycles = 19644

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  13041  13064  13077  13086  13095  13105  13117  13130  13149  13186  14450
    encaps percentiles:  15220  15248  15262  15271  15280  15289  15303  15319  15356  15442  16702
    decaps percentiles:  19556  19591  19606  19622  19633  19644  19668  19696  19738  19827  20997

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 22209
    encaps cycles = 23105
    decaps cycles = 29439

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  22111  22153  22168  22181  22193  22209  22229  22264  22332  22892  23891
    encaps percentiles:  22994  23041  23058  23074  23088  23105  23128  23160  23230  24315  38833
    decaps percentiles:  29338  29375  29390  29405  29420  29439  29464  29501  29598  30510  61061

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 31793
    encaps cycles = 33923
    decaps cycles = 42594

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  31650  31713  31741  31759  31777  31793  31820  31867  31956  33052  34458
    encaps percentiles:  33784  33831  33863  33882  33905  33923  33944  33979  34048  35182  37472
    decaps percentiles:  42387  42453  42480  42514  42540  42594  42683  42800  42906  43835  55788
```

Here are the results on the EC2 c7i instance (gcc 13.2.0):
main (75f52dc1c259d69bb4e30ef86886a9c0e01c5935):
```
./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 10718
    encaps cycles = 15185
    decaps cycles = 20685

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  10122  10398  10555  10628  10671  10718  10770  10870  11027  11306  12246
    encaps percentiles:  14388  14777  14957  15060  15140  15185  15242  15319  15441  15590  16270
    decaps percentiles:  19664  20097  20309  20449  20530  20685  20841  21026  21195  21493  22682

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 18714
    encaps cycles = 20871
    decaps cycles = 28260

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  17889  18223  18375  18508  18606  18714  18804  18931  19074  19373  19942
    encaps percentiles:  19980  20420  20605  20700  20773  20871  20959  21041  21141  21370  22024
    decaps percentiles:  27426  27727  27888  27999  28151  28260  28363  28492  28640  28843  29926

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 25033
    encaps cycles = 28978
    decaps cycles = 39337

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  23925  24472  24687  24829  24918  25033  25141  25240  25376  25588  26715
    encaps percentiles:  27844  28336  28547  28712  28855  28978  29133  29268  29435  29755  30669
    decaps percentiles:  38054  38466  38799  38996  39151  39337  39499  39688  39865  40191  41062
```

this PR (056112032b689bfa59322fdff8209390fa48736a):
```
./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 10614
    encaps cycles = 14397
    decaps cycles = 18979

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  10052  10236  10385  10480  10555  10614  10666  10707  10756  10976  11662
    encaps percentiles:  13696  13986  14128  14234  14322  14397  14475  14539  14626  14807  15460
    decaps percentiles:  18200  18539  18691  18788  18890  18979  19086  19183  19302  19438  19976

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 18471
    encaps cycles = 19655
    decaps cycles = 26197

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  17446  17969  18126  18267  18386  18471  18547  18657  18833  19133  19935
    encaps percentiles:  18701  19163  19318  19443  19546  19655  19727  19814  19985  20191  20780
    decaps percentiles:  25126  25602  25775  25950  26056  26197  26319  26426  26579  26877  27969

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 25171
    encaps cycles = 27385
    decaps cycles = 36543

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  24251  24654  24823  24959  25068  25171  25281  25408  25554  25809  26945
    encaps percentiles:  26348  26737  26936  27147  27271  27385  27517  27700  27860  28133  29019
    decaps percentiles:  35461  35813  36088  36242  36394  36543  36694  36856  37080  37408  38280
```

# Component benchmarks (trimmmed to just polyvec_compress and polyvec_decompress)

Here are the results on my Raptor Lake (gcc 14.2.1)
main (75f52dc1c259d69bb4e30ef86886a9c0e01c5935)
```
INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=808
polyvec_decompress cycles=828

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=1458
polyvec_decompress cycles=1258

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=1793
polyvec_decompress cycles=1606
```

this PR (056112032b689bfa59322fdff8209390fa48736a):
```
INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=135
polyvec_decompress cycles=81

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=202
polyvec_decompress cycles=112

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=386
polyvec_decompress cycles=163

```

Here are the results on the EC2 c7i instance (gcc 13.2.0):
main (75f52dc1c259d69bb4e30ef86886a9c0e01c5935):
```
./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt --components

INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=814
polyvec_decompress cycles=663

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=1254
polyvec_decompress cycles=995

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=1869
polyvec_decompress cycles=1408
```

this PR (056112032b689bfa59322fdff8209390fa48736a):
```
./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt --components

INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=144
polyvec_decompress cycles=121

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=216
polyvec_decompress cycles=141

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=417
polyvec_decompress cycles=233


```