@mkannwischer 2024-11-20
Here are results for clang:

**TL;DR: clang is not better at autovectorizing this.** In these results it looks like clang performs much worse on Raptor Lake vs. gcc14, and a bit better on c6i vs. gcc13. I tried to re-run the gcc14 benchmarks on Raptor Lake today and I cannot reproduce the numbers I got yesterday. Maybe I did make a mistake yesterday - but that won't matter for this PR hopefully.

# Raptor Lake (clang 18.1.8)

main (https://github.com/pq-code-package/mlkem-native/commit/75f52dc1c259d69bb4e30ef86886a9c0e01c5935)
```
CC=clang ./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 23670
    encaps cycles = 26846
    decaps cycles = 34556

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  23566  23610  23629  23642  23657  23670  23683  23700  23720  23772  26226
    encaps percentiles:  26709  26769  26792  26814  26831  26846  26864  26884  26918  26958  29406
    decaps percentiles:  34389  34458  34490  34512  34536  34556  34577  34607  34643  34699  37077

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 39521
    encaps cycles = 42353
    decaps cycles = 52863

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  39385  39435  39459  39480  39502  39521  39542  39573  39653  40222  42260
    encaps percentiles:  42203  42257  42287  42312  42332  42353  42377  42405  42482  43225  45172
    decaps percentiles:  52666  52733  52771  52805  52829  52863  52895  52944  53076  53766  55598

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 59144
    encaps cycles = 62901
    decaps cycles = 79353

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  58971  59035  59070  59095  59124  59144  59170  59203  59279  61626  64327
    encaps percentiles:  62707  62795  62823  62844  62869  62901  62932  62964  63017  65308  67772
    decaps percentiles:  79085  79190  79228  79270  79312  79353  79394  79453  79580  81775  84285
```
```
INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=948
polyvec_decompress cycles=760

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=1421
polyvec_decompress cycles=1142

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=2219
polyvec_decompress cycles=2473
```

this PR (https://github.com/pq-code-package/mlkem-native/commit/056112032b689bfa59322fdff8209390fa48736a):
```
INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 23655
    encaps cycles = 27018
    decaps cycles = 33696

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  23549  23596  23616  23629  23641  23655  23668  23681  23699  23728  26185
    encaps percentiles:  26897  26944  26970  26986  27005  27018  27034  27054  27076  27118  29615
    decaps percentiles:  33525  33591  33620  33650  33671  33696  33715  33741  33775  33845  36200

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 39444
    encaps cycles = 41144
    decaps cycles = 51520

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  39314  39361  39387  39409  39428  39444  39461  39482  39520  39888  42056
    encaps percentiles:  41007  41062  41086  41106  41124  41144  41159  41181  41215  41356  43765
    decaps percentiles:  51338  51412  51445  51470  51498  51520  51543  51584  51643  52444  54155

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 58133
    encaps cycles = 60818
    decaps cycles = 74927

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  57972  58037  58063  58087  58107  58133  58156  58194  58258  60622  63247
    encaps percentiles:  60656  60729  60755  60775  60798  60818  60846  60887  60955  63258  63749
    decaps percentiles:  74648  74757  74802  74845  74884  74927  74962  74999  75088  77309  78637
```

```
INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=333
polyvec_decompress cycles=185

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=499
polyvec_decompress cycles=259

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=914
polyvec_decompress cycles=385
```

# c7i (clang 18.1.3)
```
CC=clang ./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 9482
    encaps cycles = 13159
    decaps cycles = 17565

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:   9292   9327   9352   9388   9429   9482   9542   9602   9685   9786  10416
    encaps percentiles:  12870  12962  13001  13037  13090  13159  13239  13328  13422  13585  13982
    decaps percentiles:  17259  17320  17381  17430  17493  17565  17664  17727  17783  17852  18316

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 15944
    encaps cycles = 17761
    decaps cycles = 24158

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  15625  15680  15738  15790  15862  15944  16005  16067  16135  16279  16761
    encaps percentiles:  17407  17487  17544  17598  17675  17761  17829  17890  17959  18118  18629
    decaps percentiles:  23718  23844  23945  24049  24104  24158  24201  24254  24353  24578  25024

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 23357
    encaps cycles = 25210
    decaps cycles = 35316

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  22896  23005  23118  23237  23311  23357  23404  23453  23530  23944  25004
    encaps percentiles:  24734  24902  25022  25109  25167  25210  25268  25318  25416  25767  26203
    decaps percentiles:  34277  34743  34921  35060  35195  35316  35414  35503  35636  35863  36554
```
```
INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=506
polyvec_decompress cycles=321

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=757
polyvec_decompress cycles=488

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=1900
polyvec_decompress cycles=1108
```

this PR (https://github.com/pq-code-package/mlkem-native/commit/056112032b689bfa59322fdff8209390fa48736a):
```
CC=clang ./scripts/tests bench -c PMU  --cflags="-march=native -mtune=native -mavx2 -mbmi2 -mpopcnt -maes -flto" --opt opt

```
INFO  > Benchmark          ML-KEM-512  (native,    opt)    keypair cycles = 9454
    encaps cycles = 12845
    decaps cycles = 17037

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:   9261   9300   9328   9364   9403   9454   9498   9553   9625   9722  10208
    encaps percentiles:  12506  12575  12626  12697  12759  12845  12923  12988  13079  13262  13664
    decaps percentiles:  16681  16759  16816  16869  16953  17037  17089  17141  17199  17299  17708

INFO  > Benchmark          ML-KEM-768  (native,    opt)    keypair cycles = 15900
    encaps cycles = 17133
    decaps cycles = 23111

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  15644  15684  15720  15776  15826  15900  15992  16085  16161  16326  16791
    encaps percentiles:  16789  16851  16915  16968  17057  17133  17208  17276  17337  17486  18037
    decaps percentiles:  22660  22793  22873  22950  23057  23111  23159  23205  23278  23479  24005

INFO  > Benchmark          ML-KEM-1024 (native,    opt)    keypair cycles = 22803
    encaps cycles = 23516
    decaps cycles = 31879

           percentile      1     10     20     30     40     50     60     70     80     90     99
   keypair percentiles:  22380  22487  22554  22656  22752  22803  22864  22905  22967  23382  24115
    encaps percentiles:  23056  23197  23273  23359  23429  23516  23576  23627  23700  24004  24582
    decaps percentiles:  31309  31591  31681  31749  31820  31879  31947  32039  32150  32398  32927
```
INFO  > Benchmark Components ML-KEM-512  (native,    opt)
polyvec_compress cycles=128
polyvec_decompress cycles=85

INFO  > Benchmark Components ML-KEM-768  (native,    opt)
polyvec_compress cycles=192
polyvec_decompress cycles=114

INFO  > Benchmark Components ML-KEM-1024 (native,    opt)
polyvec_compress cycles=365
polyvec_decompress cycles=169
```